Add the scala-indent:pad-equals setting.
This setting (default t) controls whether you get extra indentation
in the cases where statements follow =. Here is an example:

    // with scala-indent:pad-equals t
    def foo(x: Int) = x match {
        case _ => 999
      }

    // with scala-indent:pad-equals nil
    def foo(x: Int) = x match {
      case _ => 999
    }

I've defaulted the setting to t to respect the old behavior, but I
strongly suspect that most users of the mode will probably want the
setting set to nil. We'll see I guess.