Highlight "var"s differently.
Currently, both var and val definitions are font-lock'd the same, which
is compeltely valid. It is rather useful to have "var"s highlighted
differently to indicate special treatment.

var highlighting is governed by the defined face scala-font-lock:var-face,
which defaults to inherit styling from 'font-lock-warning-face. The
face is customizable from the customize-face command in emacs, ie:

	M-x customize-face

Signed-off-by: Gary Pamparà <gpampara@gmail.com>