Fix bad syntax when `*/` is prefixed by an opcode.
When `*/` is prefixed by one or more members of `scala-syntax:opchar-group` (`[#:<=>@!%&*+/?\\\\^|~-]`), the `scala-syntax:propertize-special-symbols` function incorrectly marks all the characters as symbol values. This causes the `*/` not to be recognized as a comment terminator.

This commit corrects the issue by adding an exclusion to the `scala-syntax:propertize-special-symbols` function if the `match` value has a suffix of `*/`.

A concrete example can be viewed at https://github.com/scala/scala/blob/v2.11.11/src/reflect/scala/reflect/internal/Symbols.scala#L863 .