Fix exponential backoff in multiline string regexp
Since the middle of the replaced regexp was a greedy match,
it required exponential time to match (or fail to match)
a very long literal.

For instance, if we had a `"""` followed by ~700 lines of code
that didn't contain a multine string, trying to match that with the
regexp would not terminate.