I am not going to make progress just by tweaking
I have to tear down first, before I can rebuild. 20% of the algorithm gets us
80% of the way there for braces. This totally breaks whitespace again, but it
wasn't _really_ working with what I had before, and it will be much easier to
get it working once we get a bunch of stuff out of the way. So many concepts of
have shifted, and it's just too hard to read Emacs-style code which is not
declarative. More tearing down may come after this.